+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fixed bug for `upload()` methods. Now all errors are handled properly;
 * uses `crabstudio/Recaptcha` and `mirko-pagliai/cakephp-recaptcha-mailhide`
     plugins for reCAPTCHA;
+* updated for MeTools 2.13;
 * updated for dropzone 5.1;
 * updated for cakephp-database-backup 2.1.0.