+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 2.x branch
 ## 2.19 branch
+### 2.19.2
+* fixed bug in the list of active pages for each page category.
+
 ### 2.19.1
 * added configuration for reCAPTCHA (`config/recaptcha.php`);
 * fixed bug for `upload()` methods. Now all errors are handled properly;